Reaction wheel config system (#2629)
* Add ReactionWheel Patching System

* Configure parts to use ReactionWheel tool

Preserve power draw instead of specified torque.  While the torque has likely not been balanced [this is the impetus for this tool/system], the power draw likely has
Tagging parts as wheel vs CMG was not very disciplined.  Individual parts can be checked and adjusted as needed.  Applied blanket updates to existing configs.
WIP: A few parts didn't specify power draw, so complete SWAG
Unrelated: Added NEEDS guards to some legacy RemoteTech configs.  Removed a couple TWEAKSCALE modules that had no NEEDS guards.
Issue: No idea what is going on with the LazTek config, or if this mod is even still available
Removed RW from a Sputnik-alike.

* Configure Squad and VNP parts to use ReactionWheel tool

Parts more likely to be seen/used by players, separate the commit for easier inspection

* Configure RN parts to bypass ReactionWheel tools

* Mass Remove Per-Part ReactionWheel Removal

Rely on the tool to remove all reaction wheels that do not use the tool and do not flag to bypass the tool.
Results in removing ~400 part-part instances of !MODULE[ModuleReactionWheel].  That's a lot of by-hand removals and is subject to missing lots of parts.